Update the query format, type is now returned separately from basic
